% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corner.R
\name{corner}
\alias{corner}
\title{Get top left corner of a matrix}
\usage{
corner(mat, n = 6)
}
\arguments{
\item{mat}{A matrix or list object.}

\item{n}{The number of elements to show (number or rows and columns)}
}
\value{
Returns the first `n` rows and `n` columns of the matrix (top left
corner)
}
\description{
Get top left corner of a matrix
}
\examples{


}
